breakpoints with INT0 interrupt instead of timer work!
The trick is to move PC 1 word back in INT0 ISR because the PC saved on the stack points to the next instruction after the TRAPOPCODE which replaced the original instruction. So we need to move the PC back so that upon return from ISR the original instruction is executed.